chore: set up repo
Set up commit message template to make it easier for team members to
follow agreed upon specification for commits.

To make sure this template is being used, execute this command in the
root of the repository:
'git config commit.template .gitmessage'

NB! This template will be overridden if the -m flag is used when
preforming a git commit.

For more information about the template/specification, see here:
https://www.conventionalcommits.org/

Created pull request template for repo, inspired by:
https://axolo.co/blog/p/part-3-github-pull-request-template.

Refs: #11, #15